Organize codebase, switch to 'gar' for args parsing, prepare subcommands
