+++ package.json
@@ -25,11 +25,11 @@
   "dependencies": {
     "chalk": "^2.4.1",
     "fs-extra": "^7.0.0",
+    "gar": "^1.0.3",
     "jsdom": "^12.0.0",
     "moment": "^2.22.2",
     "node-fetch": "^2.2.0",
-    "query-string": "^6.1.0",
-    "yargs": "^12.0.1"
+    "query-string": "^6.1.0"
   },
   "private": true
 }
