+++ src/types.d.ts
@@ -1,0 +1,4 @@
+declare module 'gar' {
+  function parseArgv(argv: string[]): {[key: string]: boolean | string | string[]}
+  export = parseArgv
+}
