+++ tsconfig.json
@@ -5,9 +5,9 @@
     "strict": true,
     "outDir": "dist",
     "lib": ["es2015", "es2016", "es2017", "dom"],
-    "sourceMap": true
+    "sourceMap": true,
   },
   "include": [
-    "src/**/*.ts"
-  ]
+    "src/**/*.ts",
+  ],
 }
